move nx_graph back into DependencyGraph
It used to be outside of it (despite taking self as a parameter), and
would be imported explicitly by __init__. I also made it return
nx_labels instead of modifying self, since that information is
nx-specific. Lastly, since networkx is only used in nx_graph, I
moved the import in there. It can be moved out later on if more
nx-specific methods are added.